requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
langchain
faiss-cpu
langchain_community
langchain_ollama
python-dotenv
